[log] move Logger.py to PL and update all imports
