+++ TreeModelLib/BelowgroundCompetition/NetworkOGSLargeScale3D/NetworkOGSLargeScale3D.py
@@ -11,7 +11,7 @@ from os import path
 
 from TreeModelLib.BelowgroundCompetition.SimpleNetwork import SimpleNetwork
 from TreeModelLib.BelowgroundCompetition.OGSLargeScale3D import OGSLargeScale3D
-from TreeModelLib.Logger import method_logger
+from ProjectLib.Logger import method_logger
 
 
 class NetworkOGSLargeScale3D(SimpleNetwork, OGSLargeScale3D):
