+++ TreeModelLib/BelowgroundCompetition/NetworkOGSLargeScale3DExternal/NetworkOGSLargeScale3DExternal.py
@@ -9,7 +9,7 @@ from TreeModelLib.BelowgroundCompetition.OGSLargeScale3DExternal import \
      OGSLargeScale3DExternal
 from TreeModelLib.BelowgroundCompetition.NetworkOGSLargeScale3D import \
     NetworkOGSLargeScale3D
-from TreeModelLib.Logger import method_logger
+from ProjectLib.Logger import method_logger
 
 
 # Child class of OGSLargeScale3DExternal and NetworkOGSLargeScale3D to use
