+++ TreeModelLib/BelowgroundCompetition/OGSLargeScale3D/OGSLargeScale3D.py
@@ -7,7 +7,7 @@
 import numpy as np
 from TreeModelLib import TreeModel
 from TreeModelLib.BelowgroundCompetition.OGS.helpers import CellInformation
-from TreeModelLib.Logger import method_logger
+from ProjectLib.Logger import method_logger
 import vtk as vtk
 from lxml import etree
 from os import path
