+++ TreeModelLib/BelowgroundCompetition/SimpleHydro/SimpleHydro.py
@@ -7,7 +7,7 @@
 import numpy as np
 import math
 from TreeModelLib import TreeModel
-from TreeModelLib.Logger import method_logger
+from ProjectLib.Logger import method_logger
 
 
 class SimpleHydro(TreeModel):
